Rename correctly.
